lib: cogip_defs: move header to subdirectory
Move plublic headers in a subdirectory of the include directory,
with same name as module/namespace.
It makes #include statements more explicit.
