+++ examples/lidar_obstacles/lidar_obstacles.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
 // Application includes
-#include "cogip_defs.h"
+#include "cogip_defs/cogip_defs.h"
 
 void obstacle_updater_start(const pose_t *robot_state);
