+++ examples/lidar_obstacles/main.cpp
@@ -3,7 +3,7 @@
 #include "riot/thread.hpp"
 
 // Project includes
-#include "cogip_defs.h"
+#include "cogip_defs/cogip_defs.h"
 #include "shell_menu/shell_menu.hpp"
 #include "obstacles.hpp"
 #include "tracefd/tracefd.hpp"
