+++ lib/obstacles/include/obstacles.hpp
@@ -30,7 +30,7 @@
 #include "riot/mutex.hpp"
 
 // Project includes
-#include "cogip_defs.h"
+#include "cogip_defs/cogip_defs.h"
 #include "tracefd/File.hpp"
 
 #if 0 // Unused?
