+++ platforms/pf_test/lidar_obstacles.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 // Application includes
-#include "cogip_defs.h"
+#include "cogip_defs/cogip_defs.h"
 
 // Project includes
 #include "obstacles.hpp"
